requirements.txt
pytest>=8.3
pytest-html>=4.1
pytest-xdist>=3.6
pytest-playwright>=0.5
playwright>=1.44
requests>=2.32
pydantic>=2.10
python-dotenv>=1.0
allure-pytest>=2.13
faker>=25.0
